♻️ refactor: card type/time stamp 폴더 구조 변경
